# @wordpress/data:  Introduce new custom `useDispatch` react hook