--- a/packages/data/src/components/use-dispatch/index.js
+++ b/packages/data/src/components/use-dispatch/index.js
@@ -1,0 +1,2 @@
+export { default as useDispatch } from 'use-dispatch';
+export { default as useDispatchWithMap } from 'use-dispatch-with-map';