--- a/packages/data/src/components/use-dispatch/use-dispatch-with-map.js
+++ b/packages/data/src/components/use-dispatch/use-dispatch-with-map.js
@@ -4,36 +4,29 @@
 import { mapValues } from 'lodash';
 
 /**
- * WordPress dependencies
- */
-import { useMemo, useCallback } from '@wordpress/element';
-
-/**
  * Internal dependencies
  */
 import useRegistry from '../registry-provider/use-registry';
 
-const useDispatchWithMap = ( _dispatchMap, deps ) => {
+function proxyDispatch( propName, registry, ...args ) {
+	this( registry.dispatch, registry )[ propName ]( ...args );
+}
+
+const useDispatchWithMap = ( dispatchMap ) => {
 	const registry = useRegistry();
-	const dispatchMap = useCallback( _dispatchMap, deps );
-	return useMemo( () => {
-		const currentDispatchProps = dispatchMap( registry.dispatch, registry );
-		return mapValues(
-			currentDispatchProps,
-			( dispatcher, propName ) => {
-				if ( typeof dispatcher !== 'function' ) {
-					// eslint-disable-next-line no-console
-					console.warn(
-						`Property ${ propName } returned from dispatchMap in useDispatchWithMap must be a function.`
-					);
-					// only functions should be the values of properties returned from
-					// the map.
-					return () => void 0;
-				}
-				return dispatcher;
+	const currentDispatchProps = dispatchMap( registry.dispatch, registry );
+	return mapValues(
+		currentDispatchProps,
+		( dispatcher, propName ) => {
+			if ( typeof dispatcher !== 'function' ) {
+				// eslint-disable-next-line no-console
+				console.warn(
+					`Property ${ propName } returned from dispatchMap in useDispatchWithMap must be a function.`
+				);
 			}
-		);
-	}, [ registry, dispatchMap ] );
+			return proxyDispatch.bind( dispatchMap, propName, registry );
+		}
+	);
 };
 
 export default useDispatchWithMap;