--- a/packages/data/src/components/use-dispatch/use-dispatch-with-map.js
+++ b/packages/data/src/components/use-dispatch/use-dispatch-with-map.js
@@ -13,25 +13,23 @@ import { useMemo } from '@wordpress/element';
  */
 import useRegistry from '../registry-provider/use-registry';
 
-function proxyDispatch( propName, registry, ...args ) {
-	this( registry.dispatch, registry )[ propName ]( ...args );
-}
-
 const useDispatchWithMap = ( dispatchMap ) => {
 	const registry = useRegistry();
-	const currentDispatchProps = dispatchMap( registry.dispatch, registry );
-	return useMemo( () => mapValues(
-		currentDispatchProps,
-		( dispatcher, propName ) => {
-			if ( typeof dispatcher !== 'function' ) {
-				// eslint-disable-next-line no-console
-				console.warn(
-					`Property ${ propName } returned from dispatchMap in useDispatchWithMap must be a function.`
-				);
+	return useMemo( () => {
+		const currentDispatchProps = dispatchMap( registry.dispatch, registry );
+		return mapValues(
+			currentDispatchProps,
+			( dispatcher, propName ) => {
+				if ( typeof dispatcher !== 'function' ) {
+					// eslint-disable-next-line no-console
+					console.warn(
+						`Property ${ propName } returned from dispatchMap in useDispatchWithMap must be a function.`
+					);
+				}
+				return dispatcher;
 			}
-			return proxyDispatch.bind( dispatchMap, propName, registry );
-		}
-	), [ dispatchMap ] );
+		);
+	}, [ dispatchMap, registry ] );
 };
 
 export default useDispatchWithMap;