--- a/packages/data/src/components/use-dispatch/use-dispatch.js
+++ b/packages/data/src/components/use-dispatch/use-dispatch.js
@@ -1,0 +1,16 @@
+/**
+ * WordPress dependencies
+ */
+import { useMemo } from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import useRegistry from '../registry-provider/use-registry';
+
+const useDispatch = ( storeName ) => {
+	const registry = useRegistry();
+	return useMemo( () => registry.dispatch( storeName ), [ registry ] );
+};
+
+export default useDispatch;