--- a/packages/data/src/components/use-dispatch/use-dispatch.js
+++ b/packages/data/src/components/use-dispatch/use-dispatch.js
@@ -10,7 +10,11 @@ import useRegistry from '../registry-provider/use-registry';
 
 const useDispatch = ( storeName ) => {
 	const registry = useRegistry();
-	return useMemo( () => registry.dispatch( storeName ), [ registry ] );
+	return useMemo( () => {
+		return storeName === void 0 ?
+			registry.dispatch :
+			registry.dispatch( storeName );
+	}, [ registry, storeName ] );
 };
 
 export default useDispatch;