--- a/packages/data/src/components/use-dispatch/use-dispatch.js
+++ b/packages/data/src/components/use-dispatch/use-dispatch.js
@@ -1,20 +1,11 @@
 /**
- * WordPress dependencies
- */
-import { useMemo } from '@wordpress/element';
-
-/**
  * Internal dependencies
  */
 import useRegistry from '../registry-provider/use-registry';
 
 const useDispatch = ( storeName ) => {
-	const registry = useRegistry();
-	return useMemo( () => {
-		return storeName === void 0 ?
-			registry.dispatch :
-			registry.dispatch( storeName );
-	}, [ registry, storeName ] );
+	const { dispatch } = useRegistry();
+	return storeName === void 0 ? dispatch : dispatch( storeName );
 };
 
 export default useDispatch;