--- a/packages/data/src/components/use-dispatch/use-dispatch.js
+++ b/packages/data/src/components/use-dispatch/use-dispatch.js
@@ -9,7 +9,7 @@ import useRegistry from '../registry-provider/use-registry';
  * Note: The component using this hook must be within the context of a
  * RegistryProvider.
  *
- * @param {?string} storeName  Optionally provide the name of the store from
+ * @param {string} [storeName]  Optionally provide the name of the store from
  *                             which to retrieve action creators. If not
  *                             provided, the registry.dispatch function is
  *                             returned instead.