--- a/packages/data/src/components/with-dispatch/index.js
+++ b/packages/data/src/components/with-dispatch/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { pure, createHigherOrderComponent } from '@wordpress/compose';
+import { createHigherOrderComponent } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -9,13 +9,15 @@ import { pure, createHigherOrderComponent } from '@wordpress/compose';
 import { useDispatchWithMap } from '../use-dispatch';
 
 /**
- * Higher-order component used to add dispatch props using registered action creators.
+ * Higher-order component used to add dispatch props using registered action
+ * creators.
  *
- * @param {Object} mapDispatchToProps Object of prop names where value is a
- *                                    dispatch-bound action creator, or a
- *                                    function to be called with the
- *                                    component's props and returning an
- *                                    action creator.
+ * @param {Function} mapDispatchToProps A function of returning an object of
+ *                                      prop names where value is a
+ *                                      dispatch-bound action creator, or a
+ *                                      function to be called with the
+ *                                      component's props and returning an
+ *                                      action creator.
  *
  * @example
  * ```jsx
@@ -38,11 +40,22 @@ import { useDispatchWithMap } from '../use-dispatch';
  *
  * // Rendered in the application:
  * //
- * //  <SaleButton discountPercent="20">Start Sale!</SaleButton>
+ * // <SaleButton discountPercent="20">Start Sale!</SaleButton>
  * ```
  *
  * @example
- * In the majority of cases, it will be sufficient to use only two first params passed to `mapDispatchToProps` as illustrated in the previous example. However, there might be some very advanced use cases where using the `registry` object might be used as a tool to optimize the performance of your component. Using `select` function from the registry might be useful when you need to fetch some dynamic data from the store at the time when the event is fired, but at the same time, you never use it to render your component. In such scenario, you can avoid using the `withSelect` higher order component to compute such prop, which might lead to unnecessary re-renders of your component caused by its frequent value change. Keep in mind, that `mapDispatchToProps` must return an object with functions only.
+ * In the majority of cases, it will be sufficient to use only two first params
+ * passed to `mapDispatchToProps` as illustrated in the previous example.
+ * However, there might be some very advanced use cases where using the
+ * `registry` object might be used as a tool to optimize the performance of
+ * your component. Using `select` function from the registry might be useful
+ * when you need to fetch some dynamic data from the store at the time when the
+ * event is fired, but at the same time, you never use it to render your
+ * component. In such scenario, you can avoid using the `withSelect` higher
+ * order component to compute such prop, which might lead to unnecessary
+ * re-renders of your component caused by its frequent value change.
+ * Keep in mind, that `mapDispatchToProps` must return an object with functions
+ * only.
  *
  * ```jsx
  * function Button( { onClick, children } ) {
@@ -68,22 +81,19 @@ import { useDispatchWithMap } from '../use-dispatch';
  * //
  * //  <SaleButton>Start Sale!</SaleButton>
  * ```
- * _Note:_ It is important that the `mapDispatchToProps` function always returns an object with the same keys. For example, it should not contain conditions under which a different value would be returned.
  *
  * @return {Component} Enhanced component with merged dispatcher props.
  */
 const withDispatch = ( mapDispatchToProps ) => createHigherOrderComponent(
-	( WrappedComponent ) => pure(
-		( ownProps ) => {
-			const mapDispatch = ( dispatch, registry ) => mapDispatchToProps(
-				dispatch,
-				ownProps,
-				registry
-			);
-			const dispatchProps = useDispatchWithMap( mapDispatch, [] );
-			return <WrappedComponent { ...ownProps } { ...dispatchProps } />;
-		}
-	),
+	( WrappedComponent ) => ( ownProps ) => {
+		const mapDispatch = ( dispatch, registry ) => mapDispatchToProps(
+			dispatch,
+			ownProps,
+			registry
+		);
+		const dispatchProps = useDispatchWithMap( mapDispatch, [] );
+		return <WrappedComponent { ...ownProps } { ...dispatchProps } />;
+	},
 	'withDispatch'
 );