--- a/packages/data/src/components/with-dispatch/with-dispatch-new.js
+++ b/packages/data/src/components/with-dispatch/with-dispatch-new.js
@@ -1,0 +1,26 @@
+/**
+ * WordPress dependencies
+ */
+import { pure, createHigherOrderComponent } from '@wordpress/compose';
+
+/**
+ * Internal dependencies
+ */
+import { useDispatchWithMap } from '../use-dispatch';
+
+const withDispatch = ( mapDispatchToProps ) => createHigherOrderComponent(
+	( WrappedComponent ) => pure(
+		( ownProps ) => {
+			const mapDispatch = ( dispatch, registry ) => mapDispatchToProps(
+				dispatch,
+				ownProps,
+				registry
+			);
+			const dispatchProps = useDispatchWithMap( mapDispatch );
+			return <WrappedComponent { ...ownProps } { ...dispatchProps } />;
+		}
+	),
+	'withDispatch'
+);
+
+export default withDispatch;