## Description

This pull contains the work exposing a new useDispatch hook on the `@wordpress/data` package.  As per discussions in #15473 and in [#core-js slack meeting](https://make.wordpress.org/core/2019/05/29/javascript-chat-summary-may-28th-2019/), the primary api exposed for the new hook is:

```js
useDispatch( storeName: ?string ): Object|Function
```

`storeName` is optional and if provided then the action creators registered on that store are returned.  Otherwise if no `storeName` is provided, then the `registry.dispatch` function is returned.

Also in this pull I've experimented with a `useDispatchWithMap` hook to replace `withDispatch` internals with (and we could maybe _not_ expose this initially, but just implement in `withDispatch` only).  I was interested in seeing:

a. is it possible.
b. does it simplify logic
c. what impact does it have.

I have a working example in this pull and did some performance benchmarking with the following results:

**5.8.0 Tag**

Average time to load: 7834ms
Average time to DOM content load: 7073ms
Average time to type character: 86.05ms
Slowest time to type character: 246ms
Fastest time to type character: 60ms

**Current Master (at time of writing this post)**

Average time to load: 6161ms
Average time to DOM content load: 5698ms
Average time to type character: 79.06ms
Slowest time to type character: 164ms
Fastest time to type character: 57ms

**This branch**

Average time to load: 7730ms
Average time to DOM content load: 7199ms
Average time to type character: 100.29ms
Slowest time to type character: 209ms
Fastest time to type character: 68ms

With those results, I doubt we'll want to merge this with the `useDispatchWithMap` changes and instead just keep `withDispatch` as is.  Still, I'm sure there's better ways of doing that hook...

**Things to do**:

* [x] Initial feedback
* [x] Apply feedback
* [x] Add tests, fix existing tests (if needed)
* [x] Update CHANGELOG.md
* [x] Ensure documentation present
* [ ] Final review
* [ ] Release. 

## How has this been tested?
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, tests ran to see how -->
<!-- your change affects other areas of the code, etc. -->

## Screenshots <!-- if applicable -->

## Types of changes
<!-- What types of changes does your code introduce?  -->
<!-- Bug fix (non-breaking change which fixes an issue) -->
<!-- New feature (non-breaking change which adds functionality) -->
<!-- Breaking change (fix or feature that would cause existing functionality to not work as expected) -->

## Checklist:
- [ ] My code is tested.
- [ ] My code follows the WordPress code style. <!-- Check code: `npm run lint`, Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/javascript/ -->
- [ ] My code follows the accessibility standards. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/accessibility-coding-standards/ -->
- [ ] My code has proper inline documentation. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/inline-documentation-standards/javascript/ -->
- [ ] I've included developer documentation if appropriate. <!-- Handbook: https://wordpress.org/gutenberg/handbook/designers-developers/ -->