@nerrad 2019-05-29
@youknowriad, @epiqueras, and @aduth I'd love to get your feedback on this before I go further (note not in it's final state with inline docs etc).