@nerrad 2019-05-29
Performance metrics from the latest commit 7d88dd1:

Average time to load: 6410ms
Average time to DOM content load: 5893ms
Average time to type character: 92.51ms
Slowest time to type character: 192ms
Fastest time to type character: 68ms

I retested master because it seemed the performance tests were running slower on my machine and there was a bit of a difference from the original post:

Average time to load: 6831ms
Average time to DOM content load: 6220ms
Average time to type character: 105.105ms
Slowest time to type character: 266ms
Fastest time to type character: 61ms

So it actually looks like things are running a tad bit faster now :)