@nerrad 2019-05-30
The latest `useDispatchWithMap` iteration pushed doesn't affect the performance significantly and its pretty similar to master.  The performance tests are fluctuating for me (sometimes a difference of 10-15ms for average typing speed between tests) but the variance between master and this branch seems to be about the same.  Would be good to have someone else run the performance tests too.