@nerrad 2019-06-01
### New `withDispatch` (using `useDispatchWithMap` hook) vs old `withDispatch` performance tests.

Here's some performance metrics on the latest push (6946a6a). For "no Async" tests, I used the method mentioned by [@youknowriad here](https://github.com/WordPress/gutenberg/pull/15896#discussion_r289442653).  For all tests in each branch I did three runs and then averaged the values across those runs.

| Branch | avg load | avg dom load | avg type | slowest type | fastest type |
| ------ | ------ | -------- | -------| -------- | -------- |
useDispatch - with async | 7246ms | 6638ms | 101ms | 240ms | 69ms |
master - with async | 6790ms | 6208ms | 98.92ms | 227ms | 56ms |
useDispatch - no async | 9190ms | 7878ms | 280.841ms | 509ms | 223ms |
master - no async | 9035ms | 7590ms | 293ms | 631ms | 219ms |

#### Conclusions:

_master_ is faster than this branch for the "with async" tests for the average but oddly with no async (which more closely tests the changes in `withDispatch`), _this branch_ is faster!  On the whole this suggests that there are some marginal gains with the new `withDispatch` and some minor negative impact on overall performance. 

### `useDispatch` with `useMemo` and without `useMemo` performance tests.

The purpose of these tests (as suggested [here](https://github.com/WordPress/gutenberg/pull/15896#discussion_r288834516) was to see whether memoizing the returned action creators (via `useMemo`) was slower or faster than the currently implemented logic.  For the test, I created a mock component implementing `useDispatch` with a mock registry and used `react-test-renderer` to mount and unmount the component 1000 times.  I tested invoking `useDispatch` both _with_ a store name and without the store name argument. The results are below:

| type | Avg time to render with store name | Avg time to render (no store name ) |
| -----| -------------- | ------------- |
no `useMemo` | 0.048ms | 0.03ms |
with `useMemo` | 0.049ms | 0.032ms |

#### Conclusions

It appears that the difference between the two is very marginal but there _is_ an overhead cost (however minor) to implementing `useMemo` so it appears we're good to leave the current logic as is.

Overall, from a performance perspective, I think there's nothing indicating that we need to backtrack on some of these changes.  I'll continue on with covering all the changes in unit tests.